refactor(ARC-09): centralize calibrate golden-assembly lookup + document maintainer-only contract
The three calibrate commands triplicated LoadTestAssembly (reflecting over AgentEval.Tests for the embedded
golden datasets) and emitted an opaque 'Could not locate AgentEval.Tests' error — a layering inversion +
packaging trap. Extract one CalibrationGoldenAssembly.TryLocate() with a shared, actionable NotFoundMessage
that documents calibrate as a maintainer/CI-only command (run from a built solution tree, not the published
NuGet — the goldens are test fixtures). All three delegate; happy-path behavior unchanged (67 calibrate
tests green). Promoting the goldens to a shipped product dataset remains vNext.
